huge refactoring- not sure i will finish in time
i decided to approach the multiple surveys problem and it requires a lot of refactoring because the state is now completely changed.
the use effect is not updating as they should and there is a lot of debugging to do before continuing with the server side.
I'm starting a now job on may 9th, so i'm not sure i will manage to finish the project by then