refactoring back
i figuered i don't neet all this massive refactoring, i can only store the current surevy id in the redux store.
So, i refactor everything back and got back to focusing on the server side.
Right now i'm trying to implement the "add page" functionality to the backend. (still not working right, mongoose is not saving all the props in the DB, not sure way..)